// src/app/api/swagger/models/api-market-participant-offer-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiCategoryDto } from '../models/api-category-dto';
import { ApiCharacteristicValuesDto } from '../models/api-characteristic-values-dto';
import { ApiCountryDto } from '../models/api-country-dto';
import { ApiDeliveryRegionInfoDto } from '../models/api-delivery-region-info-dto';
import { ApiManufacturerCountryRegistryTypeEnum } from '../models/api-manufacturer-country-registry-type-enum';
import { ApiOfferCreatedMethodEnum } from '../models/api-offer-created-method-enum';
import { ApiOkpd2MarketDto } from '../models/api-okpd-2-market-dto';
import { ApiParticipantOfferDocumentDto } from '../models/api-participant-offer-document-dto';
import { ApiParticipantOfferSourceEnum } from '../models/api-participant-offer-source-enum';
import { ApiParticipantOfferSpecificFlagEnum } from '../models/api-participant-offer-specific-flag-enum';
import { ApiParticipantOfferStateEnum } from '../models/api-participant-offer-state-enum';
import { ApiParticipantOfferTypeEnum } from '../models/api-participant-offer-type-enum';
import { ApiPriceRegionDetailDto } from '../models/api-price-region-detail-dto';
import { ApiRegionKladrDto } from '../models/api-region-kladr-dto';
import { ApiTruTypeEnum2 } from '../models/api-tru-type-enum-2';
import { ApiUsedClassificatorTypeEnum } from '../models/api-used-classificator-type-enum';
import { ApiVatRateStateEnum } from '../models/api-vat-rate-state-enum';
export interface ApiMarketParticipantOfferDto {
  Category?: ApiCategoryDto | null;
  CharacteristicFileGuid?: string;
  CharacteristicFromUi?: boolean | null;
  CharacteristicValues?: Array<ApiCharacteristicValuesDto> | null;
  ClassificatorCode?: string | null;
  ClassificatorDescription?: string | null;
  Country?: ApiCountryDto | null;
  CreatedMethod?: ApiOfferCreatedMethodEnum;
  CurrencyCode?: string | null;
  DeliveryCost?: number | null;
  DeliveryPlace?: string | null;
  DeliveryRegionInfos?: Array<ApiDeliveryRegionInfoDto> | null;
  DeliveryRegions?: string | null;
  DuringDays?: number | null;
  ExclusiveToTrade?: boolean | null;
  ExternalSystemIdentifier?: string | null;
  ExternalSystemIdentifierNumeric?: number | null;
  Id?: number;
  IsPublicOffer?: boolean;
  KsrCode?: string | null;
  LotItemId?: number | null;
  Manufacturer?: string | null;
  ManufacturerCountryRegistryNumber?: string | null;
  ManufacturerCountryRegistryType?: ApiManufacturerCountryRegistryTypeEnum | null;
  ManufacturerCountryScore?: number | null;
  NeverExpired?: boolean;
  OfferState?: ApiParticipantOfferStateEnum | null;
  OfferTerms?: string | null;
  OfferType?: ApiParticipantOfferTypeEnum;
  OkeiCode?: string | null;
  OkeiDescription?: string | null;
  Okpd2Codes?: Array<ApiOkpd2MarketDto> | null;
  ParticipantOfferDocuments?: Array<ApiParticipantOfferDocumentDto> | null;
  ParticipantOfferSource?: ApiParticipantOfferSourceEnum;
  ParticipantOfferSpecific?: ApiParticipantOfferSpecificFlagEnum | null;
  PossibleDeliveryBySeller?: boolean;
  PossibleSelfDelivery?: boolean | null;
  PossibleSendByTransportCompany?: boolean | null;
  PriceListId?: number | null;
  PriceRegionDetails?: Array<ApiPriceRegionDetailDto> | null;
  PrintFormGuid?: string | null;
  ProductDescription?: string | null;
  ProductName?: string | null;
  PublishDateStart?: string | null;
  RecipientOrganizationInn?: string | null;
  RecipientOrganizationKpp?: string | null;
  RecipientOrganizationName?: string | null;
  RegionalProductSignChecked?: boolean;
  Regions?: Array<ApiRegionKladrDto> | null;
  RequestDateEnd?: string | null;
  RequestDateStart?: string | null;
  SelfDeliveryPlace?: string | null;
  Signature?: string | null;
  TradeId?: number | null;
  TruType?: ApiTruTypeEnum2 | null;
  UnitPrice?: number | null;
  UnitPriceInCurrency?: number | null;
  UsedClassificatorType?: ApiUsedClassificatorTypeEnum;
  VatRate?: number | null;
  VatRateState?: ApiVatRateStateEnum;
}